# src/main/java/com/eaglebank/dto/request/CreateBankAccountRequest.java
package com.eaglebank.dto.request;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.Data;

@Data
public class CreateBankAccountRequest {
    
    @NotBlank(message = "Name is required")
    @Size(max = 255, message = "Name must not exceed 255 characters")
    private String name;
    
    @NotNull(message = "Account type is required")
    private AccountType accountType;

    public enum AccountType {
        personal
    }
}